Example of soak-hw.json configuration used by the poplar hw soak test builder
Summary: This diff adds a test configuration for the poplar soak tests using hardware.  This is used in diff D24631 and will be run in the nightly hardware soak tests.  This is just an example and should be modified in the future with the actual test configurations.

Test Plan: Tested with a local poplar buildbot and slave on piputy009.  Manually put this json file in place.  The tests run and on failure attempt to post to elasticsearch.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk, saml

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk, saml

Maniphest Tasks: T18130

Differential Revision: https://phabricator.sourcevertex.net/D24635