SLIC assembler: Use of define for clarity, overprocess bug fix, couldn't resist small optimisation
Summary:
As above, discovered bug while doing:
Ref T18277

Effect is that in the <3 path with no explicit zeros and only 1 item to process then : The over read partials were input to the SLIC instruction.  There is a possibility of an exception depending on the values found when overreading.

Also, couldn't resist bundling the last instruction and sharing the store that we branch to - in each pair of macros

Test Plan: CI

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm, claudiop

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T18277

Differential Revision: https://phabricator.sourcevertex.net/D25438