Dual SLIC vertex for MK2 with partials and output half
Summary:
Dual SLIC vertex for MK2 with half precision partials and results.  Uses extra AMP unit on MK2. Implemented vertex, in assembler and integrated into the convolution library.  No change to benchmarks as this is only for Mk2 and the benchmarks are for Mk1.
For stride=1 and stride=2 cases, the improvement in speed compared to the float partials version should be that 1 inner loop / macro call is needed instead of 2.

Fixes T18277

Test Plan:
Added more tests, using the exisitng ConvPartial1x4SLIC.cpp tool, for Sim2
Added single_conv_layer tests which result in the vertex being called

Locally ran 10000 conv rand tests with partials type forced to half and selecting Sim2. All passed.  Verified that some targeted this vertex by intentionally breaking it. 20% of the default 50 tests failed with that change, so this check should therefore test the vertex well.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk, tomm, claudiop, oleksiik

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Maniphest Tasks: T18277

Differential Revision: https://phabricator.sourcevertex.net/D25573