CheckAccuracyAndCast: Avoid PUT/FP reordering
Summary:
A fix to prevent floating-point instructions from being reordered with
regards to a PUT/UPUT instruction is proving harder to fix than
anticipated. While a solution is being worked on, this commit implements
a workaround, namely to move the floating-point code into a separate non
inlinable function since functions and PUT/UPUT cannot be reordered with
regards to each other. The impact is 2 extra instructions per
CheckAccuracyWhenCast class for an single extra cycle since the call
branch is dual-issued with the UPUT instructions.

Test Plan: CI

Reviewers: #poplar_team, ranjayc, timp, samk, claudiop, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: #poplar_team, samk, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: richard, davidb

Maniphest Tasks: T19411

Differential Revision: https://phabricator.sourcevertex.net/D27535