Mixed half/float aX Plus bY vertex.
Summary:
Variants of the 'aX plus bY' vertex with 'X' and 'Y' half float; 'a' and 'b' single float.
Both Supervisor and 2D vertices are added, for both constant and tensor types for 'a' and 'b'
Addresses T13191.

The templates for aXplusbY vertices (Supervisor and  2D)  are modified to add an extra template parameter for the scaleA/B values.

The `VERTEX(supervisor)` code (in **ScaledAddSupervisor_fp.S**) is re-used for the supervisor variant, but note that we specify an atom of TWO  = 2 halfs, not 2 single floats, like elsewhere in the ScaledAdd code.

Test Plan: New test cases added for the aXplusbY codelet test, covering  supervisor and 2D vertices.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, godfrey.da.costa, timp

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, timp

Maniphest Tasks: T13191

Differential Revision: https://phabricator.sourcevertex.net/D21456