Fix exception in 1x1 convolution.
Summary:
This fixes the issue from T20763: tile exception due to accessing memory
outside the 'out' field of 'ConvPartial1x1Out' vertex.

This also adds a low level codelet test for the vertex. Note that the test is not
able to detect the issue as that is due to the exact layout of variables in
memory (see task descriptions for details).

Test Plan:
run the command:
```
 'build/poplibs/tools/single_conv_layer' '--data-type=float' '--conv-groups=1' '--batch-size=1' '--field={25,5}' '--input-channels=1' '--in-dilation={1,1}' '--padding-upper={1,2}' '--padding-lower={1,2}' '--output-channels=2' '--kernel-size={3,9}' '--kernel-dilation={1,1}' '--kernel-padding-upper={0,0}' '--kernel-padding-lower={0,0}' '--stride={1,1}' '--convolution-options={"use128BitConvUnitLoad": true, "partialsType": "float"}' '--tiles-per-ipu=16' '--ipus=1' '--device-type=Sim'
```
This causes the tile exception (or not, after the fix)

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik, godfrey.da.costa

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Subscribers: timp

Maniphest Tasks: T20763

Differential Revision: https://phabricator.sourcevertex.net/D27300