Flip group-norm back to being non-strided over channels
Summary:
This essentially undoes D8432.

Taking the implementation that undoes D8432 and adding an option to control it (by default no change, still the unconventional strided group norm).  The opinion of others in frameworks was that (optionally) passing the option to all xxxNorm.hpp functions was a more consistent API, even though the options only have an effect in a limited number of use cases.

Fixes T22030

Reviewers: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk, oleksiik

Reviewed By: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: ranjayc, timp

Maniphest Tasks: T22030

Differential Revision: https://phabricator.sourcevertex.net/D12840