Simplified the ConvProgramTree
Summary: The current control program structure allows for serial splits at every level of the hierarchy which is not supporting so complicates the code. This change simplifies the structure on the basis that only the innermost level of the hierarchy will have a Repeat loop.

Test Plan: existing tests

Reviewers: tomm, richard, saml, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, saml, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: javeda

Maniphest Tasks: T18735

Differential Revision: https://phabricator.sourcevertex.net/D27235