Changed algorithm for planning multi-convs to minimise difference between stages
Summary:
Ref T18735. Depends on D27419.

This algorithm will:
 - plan the largest (FLOPs) convolution, minimising for cycles,
 - plan the rest of the convolutions minimising the sum of the difference (lower bound clamped to 0) between the cycles for each stage of the convolution compared to the first conv.

Currently uses a fuzzy heuristic (using `std::log`) to split the tiles in such a way that the small convs still get given enough to be able to work with. This was picked on a very small sample size and so should be revisited to find a better way of doing this.

I've removed the reductionDepth constraint because i don't think it is needed. The tiles that do not do the extra stages should emit `sans` instructions to skip those syncs, and the new objective will attempt to keep the time it takes to do the entire reduction stage the same between between those tiles.

Reviewers: saml, richard, tomm, godfrey.da.costa, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: saml, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T18735

Differential Revision: https://phabricator.sourcevertex.net/D27474