New algorithm for planning multi-convs, enabled the parallel engine by default
Summary:
Ref T18735.

Implemented a new algorithm for planning multi-convs which iteratively finds plans that fit using the least amount of tiles possible.

The algorithm is as follows:
 1. plan largest first across all tiles, optimising for speed.
 2. re-plan with a % cycle backoff from fastest, optimising for tiles used.
 3. for the remaining convs from smallest to but not including 2nd largest:
   1. remove used tiles from the array
   1. plan, optimising for fitting in reference cost and then tiles used.
 4. for final conv plan, optimising to fit in reference but not limit tiles

Cycle back-off and a minimum amount of tiles per conv are configurable options with defaults picked that worked well in the limited benchmark suite there is.

The results are fairly good (most of the benchmarks go down in cycles) so have now enabled multi-convs by default.

Test Plan: existing tests and benchmarks

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, richard, godfrey.da.costa, tomm, saml

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T18735

Differential Revision: https://phabricator.sourcevertex.net/D27753