Add API to return serial splits for a dense GradW of a dynamic sparse layer
Summary:
It is expected that the dense gradW matrix multiplication may require
serialisation to fit on device. The processing to be done in each
serial split could be any functionality dictated by the sparsity
selection algorithm. As the matmul planner doesn't know the exact
functionality and the types of rearrangements that may be required,
the serial splits for the output of dense GradW pass are made available

Fixes T22688

Test Plan: CI

Reviewers: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: samk

Maniphest Tasks: T22688

Differential Revision: https://phabricator.sourcevertex.net/D28225