Basic fwd pooling vertex test, remove unused tests.
Summary:
There weren't any pooling codelet tests (We rely on the pooling_layer tests).  But there were some redundant files from test for codelets before the library got updated...

This tool tests pooling vertices and the worklist/parameter generation and is independent of the pooling planner.  This should make tests on codelets more repeatable.  The approach of the same worklist generation as is used in the library calls being used by the codelet test program is similar to that used for the SLIC vertex and avoids code repetition.

Presently only the FWD pass is tested - meaning that we can run all sum, avg vertices (bwd == fwd) and max.  we can't run maxPoolingGrad or MaxPoolingGradientScale (But that doesn't have an assembler implementation anyhow).  This is left TODO....

Ref T20946

Test Plan: Sensible looking set of tests added to CI to cover pooling vertices (No errors found)

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk, oleksiik, tomm, ranjayc

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Maniphest Tasks: T20946

Differential Revision: https://phabricator.sourcevertex.net/D27638