Use convolution to implement pooling.
Summary:
This is all off by default, under the generic named `optimizeForSpeed` option.  The generic name seems appropriate as there is the possibility in the future to optimise pooling by constraining memory allocation and making a faster pooling vertex.  So in an ideal world the same option would allow the more constraining options of convolution and pooling with constraints to be evaluated and the fastest implementation chosen.  For now it's just simple on/off though.

This diff allows a convolution to be used to implement  a pooling operation, because:
Sum pooling is like a convolution where each channel is independent and all kernel values = 1.0
Avg pooling is like a convolution where each channel is independent and  all kernel values = 1.0/(num kernel elements)

Generally this is faster than pooling, (as of comparison today, but the pooling implementation and plan could be improved).  The motivation behind this is to improve oct_conv benchmarks which use pooling.  There are some good gains there but also some regressions, see notes below.

The small changes to return cycles and memory estimates look correct and useful, see comment about using the estimates.

Fixes T22324

Test Plan:
Duplicate most pooling tests to verify correct operation and no accidental attempt to use conv where it isn't equivalent.

Change oct_conv benchmarks to use convolution to do the pooling by default

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk, oleksiik, samuelh

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk, samuelh

Subscribers: richard

Maniphest Tasks: T22324

Differential Revision: https://phabricator.sourcevertex.net/D27844