Undo explicit swapping of operands to convolution for Matmuls
Summary:
A matrix multiplication (N x M) * (M x K) now translates to a convolution
with M input channels, K output channels, field / kernel size 1 and batch
size N. I've modified the convolution code so it picks an equivalent plan
to the one it used before (except now the swapping of operands is expressed
in the transforms). For joint plans we force swap operands to be true.

To avoid a change for Matmuls that had no pass specified I've introduced a
new pass type (NONE_MATMUL) which is the same as NONE except the planner
penalises plans that don't swap operands. The fact that we penalise these
plans is somewhat arbitrary - I plan to investigate if this can be removed
but would like that to be a separate change.

Almost all the benchmarks give identical results. The one exception is the
1440x200x400. Previously we swapped the operands outside of the matmul,
unswapped them again in the plan and used the outChanFlattenDims transform
to turn the field dimension (second dimension of the RHS matrix) into
output channels. Now the second dimension of the RHS matrix maps to the
output channel axis and we just leave the operands unswapped.

The plan we pick is completely equivalent and the estimated cost is
identical, however in one case outChanFlattenDims is set and in the other
it is not set. When we create the graph for the convolution we use whether
outChanFlattenDims is set to decide whether to introduce an explicit cop
to rearrange the input operands and so we get slightly different code.
In the case of the fc_layer_1440x200x400 test the code is slightly faster
and uses slightly less memory.

Ref T16758

Reviewers: tomm, samk, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T16758

Differential Revision: https://phabricator.sourcevertex.net/D28830