T5621: Add compute set overloads to popops::fill
Summary:
Performing a reduction on an empty input with a compute set requires
populating the output tensor with a zero-like value, something which
the fill function would be well suited for if it supported a compute
set interface like the zero function.

This patch generalises the zero vertices into fill vertices and then
adds compute set overloads to the fill function and finally replaces
zero's implementation with calls to fill.

Additionally, as this patch modifies the implementation of zero, the
cycle estimator is also updated (as well as the benchmarks).

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk, tomm

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk, tomm

Maniphest Tasks: T5621

Differential Revision: https://phabricator.sourcevertex.net/D28612