Fix calculation of whether the output channels need padding
Summary:
Use the parameters after they have been transformed so we take operand
swapping into account.

Handle the fact that the combineConvGroups transform requires
rearrangement explicitly - previously we implicitly calculated a
rearrangement was required for the AMP or SLIC methods because
the number of channels for the untransformed parameters was less than
the number of input channels per group required by the vertex.

This does increase the cost estimates for the MAC method with
inChansPerGroup 1 when combineConvGroups is true, however I believe
it better reflects the true costs.

Ref: T16758

Reviewers: samk, tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: samk, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T16758

Differential Revision: https://phabricator.sourcevertex.net/D28487