Added multi_conv_layer tool
Summary:
This tool is more flexible than the octconv layer so will make it easier to write more targetted tests. Ref T21616.

After this we can add a random test generator that uses this tool and a soak test configuration as well.

Test Plan: added new test for testing a multiconv with different types (which the octconv tool cannot do)

Reviewers: tomm, saml, oleksiik, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: oleksiik, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T21616

Differential Revision: https://phabricator.sourcevertex.net/D28656