enable single inputReduce as a prelude to removing the option, and making equivalent to enabled always
Summary:
The ultimate intention is to have enableSingleInput reduce run without an option (The option this diff switched to true will be erased and act as if = true). When attempting to land that before, SDK regressions were observed.

Since then I added some convolution planner changes which locally improved the regressions (D27100), so it's time to try this again!  If there are no problems, I will remove the option altogether.

Ref T13120

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm, samk

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T13120

Differential Revision: https://phabricator.sourcevertex.net/D28703