Add support for convolutions with a batch size 0
Summary:
A convolution with a batch size of 0 is easy to define mathematically,
it simply produces an output with a batch size of 0. Supporting this
in poplibs avoids having to special case this in the caller. Add tests for
convolutions with zero input / output channels at the same time.

Depends on D28701

Fixes T23371

Reviewers: godfrey.da.costa, Eddy, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Reviewed By: godfrey.da.costa, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Maniphest Tasks: T23371

Differential Revision: https://phabricator.sourcevertex.net/D28702