T22864: Fix VARIANTS specification in some test commands
Summary:
Some of the tests didn't execute the end of their test command,
because the VARIANTS option was specified midway through the
command. Fixing this exposed two failing tests:

    Hw_default_grouped_fully_connected_half_batch_4_halffloat
    Hw_default_grouped_fully_connected_half_batch_7_halffloat

which failed due to insufficient precision in the half data type,
so this patch also works around these errors by applying the logic
already added to some tests as part of T20791 to the failing tests.

The idea is that by using initializing values of {-1, 1} with a mean
of zero, the result of various summations and multiplications tends
to stay in the high precision range of the half and float types.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml, samk, godfrey.da.costa, tomm

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml, tomm

Subscribers: Eddy

Maniphest Tasks: T22864

Differential Revision: https://phabricator.sourcevertex.net/D28767