T21728: Reorganise unit tests into directories
Summary:
This patch creates a directory tree inside tests that mirrors that of
lib, and then moves unit test code into the appropriate directory based
on what it tests. Additionally, test labels are added automatically
based on the name of the parent directory of the tests which allows one
to run all the tests for a given library without also running tests for
other libraries. A test label is also added in the multitarget test case
for the name of the tool being used, so that one can specify all tests
that use a given tool.

Note that not all tests were included in this patch as it's already very
large. So I've left out tests that didn't test library functionality
specifically and non-unit tests (e.g. component tests) for the time being.

Test Plan:
I checked that the number of poplibs tests before and after this were the
same and also that the expected labels appeared when running ctest with
the --print-labels option.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk, saml, richard, oleksiik

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk, oleksiik

Maniphest Tasks: T21728

Differential Revision: https://phabricator.sourcevertex.net/D29051