Using accurate cycle counts in pooling planning, plus fixed bug that outsize was used for exchange estimate
Summary:
Called cycle count from the planner rather than the inaccurate estimate that was used.  Also realised that the exchange estimate was optimistic as it used output field dimensions, not input so corrected that.

Changes mean that when "optimize for speed" is selected it it possible to compare the convolution and pooling implementations and choose the best one.

Reasons for regressions/general negatives noticed:
Re-arrangement costs are occasionally not estimated accurately in the pool planner - where they are large in the chosen plan that isn't reflected in the planning.  A bit like convolution, sometimes they are way out but usually they are reasonably close.
Choice between pool and conv implementations is affected by inaccuracies in both.  Usually in the re-arrangement cost.  This means we pick conv over pooling implementation in benchmarks as described in T6424.  This case was the rearrangement cost in conv was optimistic in the planning.

There's not as much changed in this as it appears:
1. (Fairly obvious) The function `getPoolingCycles` is just the original cycle estimator moved - only changes are those to deal with the worklist containing length only and the unpack overheads - Variables created at the top and used later.
2. The changes in `PoolVertices.cpp` are almost all rearrangement of existing code. The only new thing is `poolVertexCycleEstimate`.  I did a load of factoring `generateVertices` into sub functions so that I would be able to call them from within the estimator as well as in `generateVertices`.  It turned out that I could get away with only calling one - which should be faster but has the downside of not locking the estimator functionality to the generate functionality quite as much as it could be.  However I think more lightweight is good and also more modular is good.  So the factoring into functions was left in place even though it's mostly not needed.

So that leaves less code to review carefully - broad comments on the rearrangements are of course good but detailed review not so much a good use of time as it's all cut and paste into functions!

Fixes T6424

Test Plan:
CI, some benchmark improvements.

Manually compared IPUModel (Chosen plan planning estimate compute cycles) and the execution time IPUModel cycles - matched where I checked them, indicating the same parameters as the slowest tile when profiled were fed into the planner.

Some checks against benchmarks (which now pick convolution in the poplibs benchmark tests) show improvement and more logical results - See write up in T6424

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh, oleksiik, tomm

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh, oleksiik, tomm

Subscribers: tomm

Maniphest Tasks: T6424

Differential Revision: https://phabricator.sourcevertex.net/D28655