T23657: Stop update_bench.py if a benchmark has a non-zero exit code
Summary:
The process tree for `update_bench.py` looks like:

    update_bench.py
    └── test.sh (CTest)
        ├── bench.py process 1
        |   └── benchmark binary
        ├── bench.py process 2
        |   └── benchmark binary
        ├── ...
        └── bench.py process N
            └── benchmark binary

If one of the benchmark binary processes exits with a non-zero exit code
then the `subprocess.run` call in `bench.py` (with `check=True`) will
raise a `subprocess.CalledProcessError` exception. This exception will
be written to the stderr of the `bench.py` process and then cause it to
terminate. CTest will write this output (due to `--output-on-failure`)
to its stdout, which `update_bench.py` inspects.

This patch adds code to `update_bench.py` to check the output of CTest
for `subprocess.CalledProcessError` exceptions, so that in that event of
an unexpected failure of a benchmark binary, such as a segfault, the
`update_bench.py` process can stop and report the failure.

Note that not all failure cases are handled by this, such as unexpected
exceptions in bench.py or uncatchable terminations of the benchmark
binaries.

Also note that additional output from the failure is not necessarily
visible, and in the event of an unexpected failure one can re-run the
benchmark binary manually to reproduce the failure.

Test Plan:
Manually sabotaged embedding_layer.cpp so that it would
return 1 if deviceType was IpuModel, and manually sabotaged the
benchmark_results.csv file and then ran two benchmarks:

    IpuModel_default_embedding_layer
    IpuModel2_default_embedding_layer

In the IpuModel case `update_bench.py` correctly identified the error,
printed it out and exited. In the IpuModel2 case `update_bench.py`
correctly tried to update the incorrect entry in benchmark_results.csv.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik, tomm

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Maniphest Tasks: T23657

Differential Revision: https://phabricator.sourcevertex.net/D29443