Estimates correction for the init part of the pooling vertex, benchmarks
Summary:
This looked wrong, and compared to Sim is inaccurate! The error is noticeable when there are more channelGroups.

Observed:
|Channels| Sim    |IpuModel Before    |IpuModel after|
|4     |   2552   |2605               |2647|
|8    |    3627   |4513               |3769|
|16 |      5775   |8329               |6013|
|32|       10070  |15961              |10501|

I've not attempted to correct the estimate elsewhere as we are now reasonably close and my intent is to create a diff
which changes the code and therefore the cycle estimates for the rest of it.

Examples created using `POPLIBS_LOG_LEVEL=TRACE ./build/poplibs/tools/pooling_layer --channels 32 "--field={8,8}" --stride 2 --kernel-size 2  --pooling-type avg  --batch-size 1 --tiles-per-ipu=1 --data-type=half --device-type=IpuModel --inference-only --ignore-data`

Channel groups (an therefore the cause of the inaccuracy) = channels/4

Benchmarks appear worse as when planned over a whole IPU each tile usually has only 1 channel group which is like the test case with 4 channels

Ref T22299

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik, samk

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Maniphest Tasks: T22299

Differential Revision: https://phabricator.sourcevertex.net/D29449