Add optimisations to the map expressions before a graph for them is generated
Summary:
The element-wise map expressions are optimised before
fused codelets or non-fused codelets are generated
for them.

The only optimisations performed are for pow(x,y) when
y is 0.5, -0.5, 1, -1 and 2 but additional ones could
be added.

A flag to enable the optimisation is also added and is
enabled by default.

Fixes T5612

Test Plan: New tests added

Reviewers: tomm, samk, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, timp

Reviewed By: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, timp

Maniphest Tasks: T5612

Differential Revision: https://phabricator.sourcevertex.net/D29150