Increased the parallel level of the benchmarks
Summary:
Ref T23659.

Recent work into lowering the memory footprint during compilation as well as using the unstable profile format which contains size optimisations has reduced the max RSS of the benchmarks enough to be able to parallelise them more effectively.

This change updates the parallel level where possible.

To be able to set the new parallel level i ran all of the benchmarks under `/usr/bin/time` and recorded the max RSS. the buildbots have 32GB  of ram so each was parallelised either 16 times (the default), 8 times or 4 times depending on how much their peak memory was. The table of results is here: {F1011784}

Test Plan: on buildbot this reduces the time spent running the benchmarks from 26 minutes to 20

Reviewers: saml, oleksiik, Eddy, tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: oleksiik, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T23659

Differential Revision: https://phabricator.sourcevertex.net/D29618