Favour pool plans which take longer than optimum if they have a larger channel grouping
Summary:
Pooling itself can be more efficiently implemented by maximising the tiles over which the work is spread, however this impacts other layers where a larger channel grouping is important for efficient processing.  As pooling is not such a computationally expensive operation we can compromise the pooling implementation in favour of a larger channel grouping.

Ref T24273 (Where regressions were observed to become small improvements)

Reviewers: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: samk

Maniphest Tasks: T24273

Differential Revision: https://phabricator.sourcevertex.net/D29651