Histogram Supervisor and 2Dworker vertices, assembler implementation, cycle counts, test
Summary:
Vertices for half and float inputs, accepting a series of "limits" which define comparison bounds to populate "limits+1" histogram entries.  (The upper and lower histogramentries have an open bound). Limits are the same type as the data, histogram entries are always float.

Histogram entries are float as the data format can support a good numeric range of exact integer values 0,1,2,....  which won't become inaccurate with the size of tile memory.  unsigned values could be used but casting would be required - this can either be added to the vertex at a later date or done with existing poplibs functions.  Some of these choices may be part of a bigger picture, as over a whole IPU, unsigned data may be needed to accumulate the result of all of these vertices, depending on the data quantity, but in general unsigned data processing is slower.

Some assembler code bulk and slow down is created by allowing the vertices to cope with misaligned inputs, which seemed appropriate for a general purpose measurement function.

Fixes T15274

Test Plan: Added test specific to these vertices

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh, oleksiik, claudiop

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh

Maniphest Tasks: T15274

Differential Revision: https://phabricator.sourcevertex.net/D29929