Remove popops::expr::BroadcastOpType
Summary:
This is just a refactoring to remove the BroadcastOpType, no new
functionality is added. Every use of a BroadcastOpType is replaced with
the corresponding BinaryOpType. The conversion to strings, and from
BinaryOp, are removed.

The main change is for the SCALED_ADD operator.
The BroadcastOpType::SCALED_ADD, which was used only with a scale of -1 to
implement SUBTRACT, is directly replaced with BinaryOpType::SUBTRACT.
This requires modifying the codelets (both C++ and Asm code) to remove the
scale parameter from the vertex state.
In the Asm code the inner loop retain the $scale variable, which is
initialized directly with -1.0 in the asm 'preambles', instad of being read
from the vertex state.

Test Plan: run poplibs regerssion

Reviewers: timp, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: timp, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T24229

Differential Revision: https://phabricator.sourcevertex.net/D29735