Fix cycle estimation for binary ops
Summary:
Cycle estimators for binary operations not coded in assembly (ADD, MUL, SUB for
half,float) are quite imprecise. A few changes are made to improve them.

Mods to 'OpPerformanceInfo': As the compiler unrolls the loop for most cases,
the easiest way to estimate the cycles is to look at the generated assembly,
count the instructions in a loop and find the unroll factor.
So a new field is added to 'OpPerformanceInfo'. To avoid modifying too much
code, the existing fields are left as they are (including the ad-hoc adding
of cycles during the calculation)

Timing of some operations (float ATAN2, POWER, REMAINDER; int/unsigned
DIVIDE and REMAINDER) is greatly variable depending on the data. Cycle counting
is done using a random selection of values.

The tables (maps) for comparison and non-comparison operators are merged.
This gets rid of the code that was supposedly selecting the appropriate table
but had a bug. The number are also updated.

The addition of an ad-hoc number of overhead cycles per loop (2 or 5) is
refactored in a  single place instead of 4 places (for this the
'binaryOpInnerLoopCycles' function is moved from one file ot another)

Test Plan: run poplibs regression tests

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, timp

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, timp

Maniphest Tasks: T25287

Differential Revision: https://phabricator.sourcevertex.net/D31023