Poplibs histogram function with update option, plus supervisor vertex to divide work by data size
Summary:
Adds a histogram function to gather histogram data over the whole IPU.  This allows for a float or unsigned int result, with float being faster but with a potential inaccuracy when any one histogram result is large.  An update option is provided, where the result can be accumulated with a previous result.

While writing this I found that it could be slow in some cases to use a supervisor vertex when assigning each worker a limit or limits to compare to as a way of dividing work.  A second vertex was added that assigns each worker a portion of the data. There are cases where one option is much better than the other, so both are included and cycle estimates used to decide which to target.

In writing that new "divide data" vertex various options for sharing the post processing and trailing elements between workers were considered (and some implemented).  The other options took work away from worker0 and made the race where worker 0 needs to read the result from other workers either very close, or requiring nop padding.  So these methods were rejected in favour of safe execution.

Fixes T15289

Test Plan:
More codelet tests for the new codelet type.
Tests for the whole function covering each input call option.

Reviewers: samuelh, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik, samk, Richardc

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Maniphest Tasks: T15289

Differential Revision: https://phabricator.sourcevertex.net/D30261