Use dynamic slice to copy the fragment if it is beneficial
Summary:
Using a switch can result in lots of control code, especially if there
are many replicas. Copying the data to a sliceable tensor and using
dynamic slice reduces the amount of code at the cost of increasing the
amount of live tensor data at the point of the copy. Depending on the
size of the data the saving in control code may be greater than the
increase in live tensor data.

Add logic to copy the data to a sliceable tensor if that is likely to
reduce overall memory usage. Unconditionally use dynamic slice for
a standalone all-gather operation (i.e. not part of an all-reduce) as
we are free to pick the layout of the output so no rearranging copy is
necessary.

As part of this change I've pushed the point where the all-gather
result is allocated down the call stack - we now pass an optional tensor
specifying the required destination. If no destination is specified the
callee is free to allocate a tensor using whatever layout is most
efficient.

This change has little affect on the number of cycles, see below for the
how it affects an all-reduce across 16 IPUs (measured using the replicated
collectives tool).

| Size | Before | After  |
| 1K   | 81640  | 74752  |
| 4K   | 118041 | 105487 |
| 16K  | 84277  | 92560  |
| 128K | 96788  | 101218 |

It does succeed in reducing the memory requirements. For example for the
example in T24564 it reduces the amount of memory on tile 0 from 272 KB
 to 245 KBs making the example fit.

Ref T26216

Reviewers: saml, samuelh, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: saml, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T26216

Differential Revision: https://phabricator.sourcevertex.net/D31487